Initial commit Starling auditor
This has existed in some form or another since about April of this year
but hasn't been checked in anywhere until now. It contains some git
commit ids and Starling feedItemUids that are personal to me, but
those have no information content on their own as far as I know and so
should be safe to publish.