Add exceptions for unusual changes that occurred in the week of 2020-01-13.
Between 2020-01-13 and 2020-01-20 a large number of past transactions
were updated according to the updatedAt field but with no visible
changes to the contents of the transactions. During the same time, a
new query parameter for the API became mandatory, which caused some
transactions to be downloaded and integrated late. All of the updates
were checked to verify that they are of this form.