Add manually checked exceptions for many transactions sine 2022-04-03
A number of fields have suddenly become mutable upon transaction
settlement: reference, counterPartyName, counterPartyUid,
counterPartySubEntityUid. reference is the most commonly change, and
it is not great that the bank or the merchant should be allowed to
change the reference string after the fact of the transaction, but it is
not catastrophic either, and a blanket exception will probably need to
be added for that. The latter three, on the other hand, seem
problematic. It does not make sense to me that a merchant's name or
unique identity should be allowed to change after the fact.